forms-flow-data-layer/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
astroid==3.3.8
beanie==1.29.0
black==25.1.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
coverage==7.6.12
dill==0.3.9
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.11
fastapi-cli==0.0.7
flake8==7.1.2
graphql-core==3.2.6
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
isort==6.0.1
Jinja2==3.1.5
lazy-model==0.2.0
logging==0.4.9.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
motor==3.7.0
mypy-extensions==1.0.0
orjson==3.10.15
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pycodestyle==2.12.1
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==3.2.0
Pygments==2.19.1
pylint==3.3.4
pymongo==4.11.2
pytest==8.3.5
pytest-asyncio==0.25.3
pytest-cov==6.0.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
redis==5.2.1
rich==13.9.4
rich-toolkit==0.13.2
setuptools==75.8.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.0
strawberry-graphql==0.262.0
toml==0.10.2
tomlkit==0.13.2
typer==0.15.2
typing_extensions==4.12.2
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0
